+++ R/CFCalendar360.R
@@ -29,7 +29,7 @@ CFCalendar360 <- R6::R6Class("CFCalendar360",
     #'   with `TRUE` for valid days and `FALSE` for invalid days, or `NA` where
     #'   the row in argument `ymd` has `NA` values.
     valid_days = function(ymd) {
-      ymd$year & ymd$month >= 1L & ymd$month <= 12L & ymd$day >= 1L & ymd$day <= 30L
+      !is.na(ymd$year) & ymd$month >= 1L & ymd$month <= 12L & ymd$day >= 1L & ymd$day <= 30L
     },
 
     #' @description Determine the number of days in the month of the calendar.