+++ R/CFCalendarProleptic.R
@@ -39,7 +39,7 @@ CFCalendarProleptic <- R6::R6Class("CFCalendarProleptic",
     #'   with `TRUE` for valid days and `FALSE` for invalid days, or `NA` where
     #'   the row in argument `ymd` has `NA` values.
     valid_days = function(ymd) {
-      ymd$year & ymd$month >= 1L & ymd$month <= 12L & ymd$day >= 1L &
+      !is.na(ymd$year) & ymd$month >= 1L & ymd$month <= 12L & ymd$day >= 1L &
       ifelse(self$leap_year(ymd$year),
         ymd$day <= c(31, 29, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)[ymd$month],
         ymd$day <= c(31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)[ymd$month])