+++ tests/testthat/test-CFtime.R
@@ -120,7 +120,11 @@ test_that("test all variants of creating a CFtime object and useful functions", 
   expect_true(slice(t, c("2022-01-01", "2023-01-02"))[1]) # early extreme before timeseries
   expect_true(all(!slice(t, c("2023-02-01", "2023-03-01")))) # both extremes outside time series
   expect_equal(sum(slice(t, c("2023-01-01 00:00", "2023-01-01 04:00", "2023-01-02 00:00"))), 24)
-  expect_equal(sum(slice(t, c("2023-01-01 00:00", "2023-01-01 04:00", "2023-01-02 00:00"), TRUE)), 25)
+  t_slice <- slice(t, c("2023-01-01 00:00", "2023-01-01 04:00", "2023-01-02 00:00"), TRUE)
+  expect_equal(sum(t_slice), 25)
+  new_t <- attr(t_slice, "CFTime")
+  expect_identical(new_t, t_sub <- t$subset(1:25))
+  expect_identical(t_sub, t_sub$copy())
 
   t$bounds <- TRUE
   s <- t$slice(c("2023-01-01", "2023-05-01"))